@@ -1,6 +1,6 @@
 {
   "name": "WalletIN2",
-  "version": "2.3.0",
+  "version": "1.1.1",
   "author": "Ionic Framework",
   "homepage": "https://ionicframework.com/",
   "scripts": {
@@ -9,7 +9,7 @@
     "build": "ng build",
     "watch": "ng build --watch --configuration development",
     "test": "ng test",
-    "test:ci": "ng test --browsers=Headless --no-watch", 
+    "test:ci": "ng test --browsers=Headless --no-watch",
     "lint": "ng lint",
     "eje": "copy customTheme customTheme2"
   },
@@ -35,7 +35,7 @@
     "@zxing/library": "^0.19.2",
     "@zxing/ngx-scanner": "^3.8.1",
     "angular-auth-oidc-client": "^16.0.1",
-    "angular2-qrcode": "^2.0.3",
+    "angularx-qrcode": "^15.0.1",
     "ionicons": "^7.0.0",
     "ngx-indexed-db": "^16.0.0",
     "rxjs": "~7.5.0",
@@ -44,11 +44,11 @@
   },
   "devDependencies": {
     "@angular-devkit/build-angular": "^15.0.0",
-    "@angular-eslint/builder": "^15.0.0",
-    "@angular-eslint/eslint-plugin": "^15.0.0",
-    "@angular-eslint/eslint-plugin-template": "^15.0.0",
-    "@angular-eslint/schematics": "^15.0.0",
-    "@angular-eslint/template-parser": "^15.0.0",
+    "@angular-eslint/builder": "15.2.1",
+    "@angular-eslint/eslint-plugin": "15.2.1",
+    "@angular-eslint/eslint-plugin-template": "15.2.1",
+    "@angular-eslint/schematics": "15.2.1",
+    "@angular-eslint/template-parser": "15.2.1",
     "@angular/cli": "^15.0.0",
     "@angular/compiler": "^15.0.0",
     "@angular/compiler-cli": "^15.0.0",
@@ -58,12 +58,9 @@
     "@ionic/angular-toolkit": "^9.0.0",
     "@types/jasmine": "~4.0.0",
     "@types/node": "^12.11.1",
-    "@typescript-eslint/eslint-plugin": "5.3.0",
-    "@typescript-eslint/parser": "5.3.0",
-    "eslint": "^7.6.0",
-    "eslint-plugin-import": "2.22.1",
-    "eslint-plugin-jsdoc": "30.7.6",
-    "eslint-plugin-prefer-arrow": "1.2.2",
+    "@typescript-eslint/eslint-plugin": "5.48.2",
+    "@typescript-eslint/parser": "5.48.2",
+    "eslint": "^8.33.0",
     "jasmine-core": "~4.3.0",
     "jasmine-spec-reporter": "~5.0.0",
     "jest": "^29.7.0",
