@@ -1,14 +1,11 @@
-import {Component, Input} from '@angular/core';
-import {CommonModule} from '@angular/common';
-import {FormsModule} from '@angular/forms';
-import {IonicModule, PopoverController} from '@ionic/angular';
-import {BarcodeScannerComponent} from '../../components/barcode-scanner/barcode-scanner.component';
-import {CameraService} from 'src/app/services/camera.service';
-import {AuthenticationService} from 'src/app/services/authentication.service';
-import {LogoutPage } from '../logout/logout.page';
-import {TranslateModule} from '@ngx-translate/core';
-import { Router,RouterModule} from '@angular/router';
-
+import { Component, Input, inject } from '@angular/core';
+import { CommonModule } from '@angular/common';
+import { FormsModule } from '@angular/forms';
+import { IonicModule } from '@ionic/angular';
+import { BarcodeScannerComponent } from '../../components/barcode-scanner/barcode-scanner.component';
+import { CameraService } from 'src/app/services/camera.service';
+import { TranslateModule } from '@ngx-translate/core';
+import { distinctUntilChanged, map, shareReplay } from 'rxjs';
 @Component({
   selector: 'app-camera-selector',
   templateUrl: './camera-selector.page.html',
@@ -18,33 +15,32 @@ import { Router,RouterModule} from '@angular/router';
     IonicModule,
     CommonModule,
     FormsModule,
-    BarcodeScannerComponent,
     TranslateModule,
-    RouterModule
+    BarcodeScannerComponent,
   ],
 })
+// eslint-disable-next-line @angular-eslint/component-class-suffix
 export class CameraSelectorPage {
-  selectedDevice: string = '';
-  userName: string = '';
-  @Input() availableDevices: MediaDeviceInfo[] = []; 
-  constructor( private cameraService:CameraService,
-    private router: Router,
-    private authenticationService: AuthenticationService,
-    private popoverController: PopoverController,
-  ) {}
+  @Input() public availableDevices: MediaDeviceInfo[] = [];
+  public cameraService = inject(CameraService);
 
-  ngOnInit() {
-    this.userName = this.authenticationService.getName();
-  }
+  public selectedDevice = this.cameraService.navCamera$.pipe(
+    map((device) => {
+      return device.deviceId;
+    }),
+    distinctUntilChanged(),
+    shareReplay(1)
+  );
 
-  availableDevicesEmit(devices: MediaDeviceInfo[]) {
+  public availableDevicesEmit(devices: MediaDeviceInfo[]) {
     this.availableDevices = devices;
   }
 
-  onDeviceSelectChange(selected: string) {
-    this.selectedDevice = selected;
+  public onDeviceSelectChange(selected: string) {
     if (selected != '') {
-      const device: MediaDeviceInfo | undefined = this.availableDevices.find((x) => x.deviceId === selected);
+      const device: MediaDeviceInfo | undefined = this.availableDevices.find(
+        (x) => x.deviceId === selected
+      );
       if (device != undefined) {
         this.cameraService.changeCamera(device);
       }
@@ -52,23 +48,4 @@ export class CameraSelectorPage {
       this.cameraService.noCamera();
     }
   }
-
-  logout(){
-    this.authenticationService.logout().subscribe(()=>{
-      this.router.navigate(['/login'], {})
-
-    });
-  }
-
-  async openPopover(ev: any) {
-    const popover = await this.popoverController.create({
-      component: LogoutPage, 
-      event: ev,
-      translucent: true,
-      cssClass: 'custom-popover'
-    });
-  
-    await popover.present();
-  }
-
 }
