@@ -1,35 +1,39 @@
-import { ComponentFixture, TestBed, async, waitForAsync, fakeAsync } from '@angular/core/testing';
-import { IonicModule } from '@ionic/angular';
+import { ComponentFixture, TestBed, fakeAsync, tick, waitForAsync } from '@angular/core/testing';
+import { AlertController, IonicModule } from '@ionic/angular';
 import { TranslateModule } from '@ngx-translate/core';
 import { RouterTestingModule } from '@angular/router/testing';
-import { ActivatedRoute } from '@angular/router';
+import { ActivatedRoute, Router } from '@angular/router';
 import { of } from 'rxjs';
 import { CredentialsPage } from './credentials.page';
 import { WalletService } from 'src/app/services/wallet.service';
 import { WebsocketService } from 'src/app/services/websocket.service';
 import { DataService } from 'src/app/services/data.service';
-import { environment } from 'src/environments/environment';
 import { AuthenticationService } from 'src/app/services/authentication.service';
 import { HttpClientTestingModule, HttpTestingController } from '@angular/common/http/testing';
+import { VerifiableCredential } from 'src/app/interfaces/verifiable-credential';
 
 describe('CredentialsPage', () => {
   let component: CredentialsPage;
   let fixture: ComponentFixture<CredentialsPage>;
   let walletServiceSpy: jasmine.SpyObj<WalletService>;
   let websocketServiceSpy: jasmine.SpyObj<WebsocketService>;
-  let authServiceSpy: jasmine.SpyObj<AuthenticationService>;
   let httpTestingController: HttpTestingController;
-  const mockedStats = '';
+  let router: Router;
+
+  const TIME_IN_MS = 10000;
 
   beforeEach(waitForAsync(() => {
-    walletServiceSpy = jasmine.createSpyObj('WalletService', ['requestCredential']);
+    walletServiceSpy = jasmine.createSpyObj('WalletService', ['getAllVCs', 'requestCredential', 'deleteVC', 'executeContent']);
     websocketServiceSpy = jasmine.createSpyObj('WebsocketService', ['connect']);
     const dataServiceSpyObj = jasmine.createSpyObj('DataService', ['listenDid']);
     const authServiceSpyObj = jasmine.createSpyObj('AuthenticationService', ['getName']);
-    dataServiceSpyObj.listenDid.and.returnValue(of(''));
+    walletServiceSpy.requestCredential.and.returnValue(of({} as any));
+
+    dataServiceSpyObj.listenDid.and.returnValue(of('someDidValue'));
+    walletServiceSpy.executeContent.and.returnValue(of({} as any));
+    walletServiceSpy.requestCredential.and.returnValue(of({} as any));
 
     TestBed.configureTestingModule({
-      //declarations: [CredentialsPage],
       imports: [
         IonicModule.forRoot(),
         TranslateModule.forRoot(),
@@ -51,39 +55,230 @@ describe('CredentialsPage', () => {
     }).compileComponents();
 
     httpTestingController = TestBed.inject(HttpTestingController);
-    //authServiceSpy = TestBed.inject(AuthenticationService) as jasmine.SpyObj<AuthenticationService>;
-    //dataServiceSpy = TestBed.inject(DataService) as jasmine.SpyObj<DataService>;
     fixture = TestBed.createComponent(CredentialsPage);
     component = fixture.componentInstance;
-    }));
+    router = TestBed.inject(Router);
+  }));
 
 
   it('should create', () => {
     expect(component).toBeTruthy();
   });
 
-  /*it('should call generateCred and refresh when requestCredential is successful', waitForAsync(() => {
-    const executionResponseMock = 'mockExecutionResponse';
-    walletServiceSpy.requestCredential.and.returnValue(of(executionResponseMock));
+  it('should enable scan mode when scan is called', () => {
+    component.scan();
+    expect(component.toggleScan).toBeTrue();
+    expect(component.show_qr).toBeTrue();
+    expect(component.ebsiFlag).toBeFalse();
+  });
 
+  it('should copy "did-text" to clipboard when copyToClipboard is called with "did-text"', async () => {
+    spyOn(navigator.clipboard, 'writeText');
+
+    const didText = 'DID: exampleDid';
+    const expectedClipboardContent = 'exampleDid';
+    document.body.innerHTML = `<div id="did-text">${didText}</div>`;
+
+    await component.copyToClipboard('did-text');
+
+    expect(navigator.clipboard.writeText).toHaveBeenCalledWith(expectedClipboardContent);
+  });
+
+  it('should handle error gracefully if clipboard API fails', async () => {
+    spyOn(navigator.clipboard, 'writeText').and.rejectWith('Test error');
+    spyOn(console, 'error');
+
+    const didTextContent = 'DID: exampleDid';
+    document.body.innerHTML = `<div id="did-text">${didTextContent}</div>`;
+
+    await component.copyToClipboard('did-text');
+
+    expect(console.error).toHaveBeenCalledWith('Error al copiar texto al portapapeles:', 'Test error');
+  });
+
+  it('should generate credential when generateCred is called', () => {
+    const mockCredentialOfferUri = 'mockCredentialOfferUri';
+
+    walletServiceSpy.requestCredential.and.returnValue(of());
+
+    component.credentialOfferUri = mockCredentialOfferUri;
     component.generateCred();
 
-    expect(websocketServiceSpy.connect).toHaveBeenCalledWith(
-      environment.websocket.url + environment.websocket.uri
-    );
-    expect(walletServiceSpy.requestCredential).toHaveBeenCalledWith('mockCredentialOfferUri');
-    expect(component.refresh).toHaveBeenCalled();
+    expect(walletServiceSpy.requestCredential).toHaveBeenCalledWith(mockCredentialOfferUri);
+  });
+
+  it('should update the credential list when refresh is called', fakeAsync(() => {
+    const mockCredList: VerifiableCredential[] = [
+      {
+        "@context": ["https://www.w3.org/ns/credentials/v2"],
+        "id": "1",
+        "type": ["VerifiableCredential", "SomeType"],
+        "issuer": {
+          "id": "issuerId1"
+        },
+        "issuanceDate": "2021-01-01T00:00:00Z",
+        "validFrom": "2021-01-01T00:00:00Z",
+        "expirationDate": "2025-12-31T23:59:59Z",
+        "credentialSubject": {
+          "mandate": {
+            "id": "mandateId1",
+            "mandator": {
+              "organizationIdentifier": "orgId1",
+              "commonName": "Common Name",
+              "emailAddress": "email@example.com",
+              "serialNumber": "serialNumber1",
+              "organization": "Organization Name",
+              "country": "Country"
+            },
+            "mandatee": {
+              "id": "personId1",
+              "first_name": "First",
+              "last_name": "Last",
+              "gender": "Gender",
+              "email": "email@example.com",
+              "mobile_phone": "+1234567890"
+            },
+            "power": [
+              {
+                "id": "powerId1",
+                "tmf_type": "Domain",
+                "tmf_domain": ["SomeDomain"],
+                "tmf_function": "SomeFunction",
+                "tmf_action": ["SomeAction"]
+              }
+            ],
+            "life_span": {
+              "start_date_time": "2021-01-01T00:00:00Z",
+              "end_date_time": "2025-12-31T23:59:59Z"
+            }
+          }
+        }
+      }
+    ];
+
+    walletServiceSpy.getAllVCs.and.returnValue(of(mockCredList));
+
+    component.refresh();
+    tick();
+
+    expect(component.credList).toEqual(mockCredList.reverse());
   }));
 
-  it('should handle error when requestCredential fails', waitForAsync(() => {
-    walletServiceSpy.requestCredential.and.throwError('Error');
 
-    component.generateCred();
+  it('vcDelete should call deleteVC on the wallet service with the correct ID and refresh the list', () => {
+    const testCredential: VerifiableCredential = {
+      "@context": ["https://www.w3.org/ns/credentials/v2"],
+      "id": "testCredentialId",
+      "type": ["VerifiableCredential", "SomeType"],
+      "issuer": {
+        "id": "issuerId1"
+      },
+      "issuanceDate": "2024-04-02T09:23:22.637345122Z",
+      "validFrom": "2024-04-02T09:23:22.637345122Z",
+      "expirationDate": "2025-04-02T09:23:22.637345122Z",
+      "credentialSubject": {
+        "mandate": {
+          "id": "mandateId1",
+          "mandator": {
+            "organizationIdentifier": "orgId1",
+            "commonName": "Common Name",
+            "emailAddress": "email@example.com",
+            "serialNumber": "serialNumber1",
+            "organization": "Organization Name",
+            "country": "Country"
+          },
+          "mandatee": {
+            "id": "personId1",
+            "first_name": "First",
+            "last_name": "Last",
+            "gender": "Gender",
+            "email": "email@example.com",
+            "mobile_phone": "+1234567890"
+          },
+          "power": [
+            {
+              "id": "powerId1",
+              "tmf_type": "Domain",
+              "tmf_domain": ["SomeDomain"],
+              "tmf_function": "SomeFunction",
+              "tmf_action": ["SomeAction"]
+            }
+          ],
+          "life_span": {
+            "start_date_time": "2024-04-02T09:23:22.637345122Z",
+            "end_date_time": "2025-04-02T09:23:22.637345122Z"
+          }
+        }
+      }
+    };
 
-    expect(websocketServiceSpy.connect).toHaveBeenCalledWith(
-      environment.websocket.url + environment.websocket.uri
-    );
-    expect(walletServiceSpy.requestCredential).toHaveBeenCalledWith('mockCredentialOfferUri');
-    expect(component.refresh).not.toHaveBeenCalled();
-  }));*/
+    walletServiceSpy.deleteVC.and.returnValue(of('Success'));
+    spyOn(component, 'refresh');
+    component.vcDelete(testCredential);
+    expect(walletServiceSpy.deleteVC).toHaveBeenCalledWith(testCredential.id);
+    expect(component.refresh).toHaveBeenCalled();
+  });
+
+
+  it('ngOnInit should initialize component properties and call refresh', () => {
+    spyOn(component, 'refresh');
+    component.ngOnInit();
+    expect(component.scaned_cred).toBeFalse();
+    expect(component.refresh).toHaveBeenCalled();
+  });
+
+  it('qrCodeEmit should process QR code and potentially change state or call services', fakeAsync(() => {
+    spyOn(router, 'navigate');
+    const testQrCode = "someTestQrCode";
+    component.qrCodeEmit(testQrCode);
+    tick();
+
+    expect(router.navigate).toHaveBeenCalledWith(['/tabs/vc-selector/'], { queryParams: { executionResponse: JSON.stringify({}) } });
+  }));
+
+  it('should handle alert Cancel correctly', fakeAsync(() => {
+    let alertController = TestBed.inject(AlertController);
+    spyOn(alertController, 'create').and.returnValue(Promise.resolve({
+      present: () => Promise.resolve(),
+      buttons: [
+        {
+          text: 'Cancelar',
+          role: 'cancel',
+          handler: () => {}
+        },
+        {
+          text: 'Aceptar',
+          handler: () => {}
+        }
+      ]
+    } as any));
+
+    component.credentialClick();
+    tick();
+  }));
+
+  it('should handle alert Accept correctly', fakeAsync(() => {
+    let alertController = TestBed.inject(AlertController);
+    spyOn(alertController, 'create').and.returnValue(Promise.resolve({
+      present: () => Promise.resolve(),
+      buttons: [
+        {
+          text: 'Cancelar',
+          role: 'cancel',
+          handler: () => {}
+        },
+        {
+          text: 'Aceptar',
+          handler: () => {
+          }
+        }
+      ]
+    } as any));
+
+    component.credentialClick();
+    tick();
+  }));
+
+
+
 });
