@@ -1,48 +1,38 @@
-import { ComponentFixture, TestBed, waitForAsync } from '@angular/core/testing';
-import { IonicModule, PopoverController } from '@ionic/angular';
+import { ComponentFixture, TestBed } from '@angular/core/testing';
+import { IonicModule } from '@ionic/angular';
+import { RouterTestingModule } from '@angular/router/testing';
+import { TranslateModule } from '@ngx-translate/core';
 import { ActivatedRoute, Router } from '@angular/router';
-import { FormsModule } from '@angular/forms';
-import { TranslateModule, TranslateService } from '@ngx-translate/core';
-import { of } from 'rxjs';
-
 import { HomePage } from './home.page';
-import { WalletService } from 'src/app/services/wallet.service';
-import { AuthenticationService } from 'src/app/services/authentication.service';
+import { BehaviorSubject } from 'rxjs';
 
 describe('HomePage', () => {
   let component: HomePage;
   let fixture: ComponentFixture<HomePage>;
-  let routerSpy: jasmine.SpyObj<Router>;
-  let walletServiceSpy: jasmine.SpyObj<WalletService>;
-  let authServiceSpy: jasmine.SpyObj<AuthenticationService>;
-  let popoverControllerSpy: jasmine.SpyObj<PopoverController>;
-  let activatedRoute: any; // You might need to define a spy for ActivatedRoute if necessary
+  let router: Router;
+  let route: ActivatedRoute;
 
-  beforeEach(waitForAsync(() => {
-    const routerSpyObj = jasmine.createSpyObj('Router', ['navigate']);
-    const walletServiceSpyObj = jasmine.createSpyObj('WalletService', ['executeContent']);
-    const authServiceSpyObj = jasmine.createSpyObj('AuthenticationService', ['getName', 'logout']);
-    const popoverSpyObj = jasmine.createSpyObj('PopoverController', ['create', 'present']);
-
-    TestBed.configureTestingModule({
-      //declarations: [HomePage],
-      imports: [IonicModule.forRoot(), FormsModule, TranslateModule.forRoot()],
-      providers: [
-        { provide: Router, useValue: routerSpyObj },
-        { provide: WalletService, useValue: walletServiceSpyObj },
-        { provide: AuthenticationService, useValue: authServiceSpyObj },
-        { provide: PopoverController, useValue: popoverSpyObj },
-        { provide: ActivatedRoute, useValue: { queryParams: of({}) } }, // Mock ActivatedRoute as needed
-        TranslateService, // Add if needed
+  beforeEach(async () => {
+    await TestBed.configureTestingModule({
+      imports: [
+        IonicModule.forRoot(),
+        RouterTestingModule.withRoutes([]),
+        TranslateModule.forRoot()
       ],
+      providers: [
+        {
+          provide: ActivatedRoute,
+          useValue: {
+            queryParams: new BehaviorSubject({ credential_offer_uri: 'someUri' })
+          }
+        }
+      ]
     }).compileComponents();
 
-    authServiceSpy = TestBed.inject(AuthenticationService) as jasmine.SpyObj<AuthenticationService>;
-  }));
-
-  beforeEach(() => {
     fixture = TestBed.createComponent(HomePage);
     component = fixture.componentInstance;
+    router = TestBed.inject(Router);
+    route = TestBed.inject(ActivatedRoute);
     fixture.detectChanges();
   });
 
@@ -50,17 +40,21 @@ describe('HomePage', () => {
     expect(component).toBeTruthy();
   });
 
+  it('should navigate based on queryParams on init', async () => {
+    const navigateSpy = spyOn(router, 'navigate');
 
+    (route.queryParams as BehaviorSubject<any>).next({ credential_offer_uri: 'newUri' });
 
-  it('should set userName from AuthenticationService on ngOnInit', () => {
-    const userName = 'JohnDoe';
-    authServiceSpy.getName.and.returnValue(userName);
+    await fixture.whenStable();
+    fixture.detectChanges();
 
-    component.ngOnInit();
-
-    expect(component.userName).toBe(userName);
+    expect(navigateSpy).toHaveBeenCalledWith(['/tabs/credentials'], { queryParams: { credentialOfferUri: 'newUri' } });
   });
 
 
-
+  it('startScan should navigate with specific queryParams', () => {
+    const navigateSpy = spyOn(router, 'navigate');
+    component.startScan();
+    expect(navigateSpy).toHaveBeenCalledWith(['/tabs/credentials/'], { queryParams: { toggleScan: true, from: 'home', show_qr: true } });
+  });
 });
