@@ -1,22 +1,20 @@
 import { Component, OnInit } from '@angular/core';
 import { CommonModule } from '@angular/common';
 import { FormsModule } from '@angular/forms';
-import { IonicModule, PopoverController } from '@ionic/angular';
-import { StorageService } from 'src/app/services/storage.service';
-import { QRCodeModule } from 'angular2-qrcode';
-import { VCReply, WalletService } from 'src/app/services/wallet.service';
+import { AlertController, IonicModule } from '@ionic/angular';
+import { QRCodeModule } from 'angularx-qrcode';
+import { WalletService } from 'src/app/services/wallet.service';
 import { ActivatedRoute, Router } from '@angular/router';
 import { TranslateModule, TranslateService } from '@ngx-translate/core';
-import { AuthenticationService } from 'src/app/services/authentication.service';
 import { VcViewComponent } from '../../components/vc-view/vc-view.component';
-import { LogoutPage } from '../logout/logout.page';
+import { VCReply } from 'src/app/interfaces/verifiable-credential-reply';
+import { VerifiableCredential } from 'src/app/interfaces/verifiable-credential';
 
 @Component({
   selector: 'app-vc-selector',
   templateUrl: './vc-selector.page.html',
   styleUrls: ['./vc-selector.page.scss'],
   standalone: true,
-  providers: [StorageService],
   imports: [
     IonicModule,
     CommonModule,
@@ -26,80 +24,24 @@ import { LogoutPage } from '../logout/logout.page';
     VcViewComponent,
   ],
 })
+// eslint-disable-next-line @angular-eslint/component-class-suffix
 export class VcSelectorPage implements OnInit {
-  isClick: boolean[] = [];
-  selCredList: any[] = [];
-  credList: any[] = [];
-  credDataList: any[] = [];
-  size: number = 300;
-  executionResponse: any;
-  userName: string = '';
+  public isClick: boolean[] = [];
+  public selCredList: VerifiableCredential[] = [];
+  public credList: VerifiableCredential[] = [];
+  public credDataList: VerifiableCredential[] = [];
+  public size = 300;
+  // eslint-disable-next-line @typescript-eslint/no-explicit-any
+  public executionResponse: any;
+  public userName = '';
+  public isAlertOpen = false;
+  public alertButtons = ['OK'];
 
-  _VCReply: VCReply = {
+  public _VCReply: VCReply = {
     selectedVcList: [],
     state: '',
     redirectUri: '',
   };
-  test: any = `<img src="../assets/icon/Tick/checkmark-verd.png" alt="g-maps" style="border-radius: 2px">`;
-  constructor(
-    private router: Router,
-    private walletService: WalletService,
-    private route: ActivatedRoute,
-    private authenticationService: AuthenticationService,
-    private popoverController: PopoverController,
-    public translate: TranslateService
-  ) {
-    this.route.queryParams.subscribe((params) => {
-      this.executionResponse = JSON.parse(params['executionResponse']);
-      this._VCReply.redirectUri = this.executionResponse['redirectUri'];
-      this._VCReply.state = this.executionResponse['state'];
-    });
-  }
-
-  ngOnInit() {
-    this.userName = this.authenticationService.getName();
-
-    this.credList = this.executionResponse['selectableVcList'];
-    this.credList.forEach((credential) => {
-      this.isClick.push(false);
-    });
-  }
-
-  isClicked(index: number) {
-    return this.isClick[index];
-  }
-
-  async openPopover(ev: any) {
-    const popover = await this.popoverController.create({
-      component: LogoutPage,
-      event: ev,
-      translucent: true,
-    });
-
-    await popover.present();
-  }
-
-  selectCred(cred: any, index: number) {
-    this.selCredList.push(cred);
-    this.isClick[index] = !this.isClick[index];
-  }
-  sendCred(cred: any) {
-    this.selCredList.push(cred);
-    this._VCReply.selectedVcList = this.selCredList;
-    this.walletService.executeVC(this._VCReply).subscribe({
-      next: (authenticationResponse) => {
-        this.isAlertOpen = true;
-        this.isAlertOpen = true;
-      },
-      error: (err) => {
-        let TIME_IN_MS = 1500;
-        setTimeout(() => {
-          this.isAlertOpenFail = false;
-        }, TIME_IN_MS);
-        this.isAlertOpenFail = true;
-      },
-    });
-  }
 
   public closeButton = [
     {
@@ -112,12 +54,72 @@ export class VcSelectorPage implements OnInit {
     },
   ];
 
-  isAlertOpenFail = false;
+  public constructor(
+    private router: Router,
+    private walletService: WalletService,
+    private route: ActivatedRoute,
+    public translate: TranslateService,
+    private alertController: AlertController
+  ) {
+    this.route.queryParams.subscribe((params) => {
+      this.executionResponse = JSON.parse(params['executionResponse']);
+      this._VCReply.redirectUri = this.executionResponse['redirectUri'];
+      this._VCReply.state = this.executionResponse['state'];
+    });
+  }
 
-  isAlertOpen = false;
-  public alertButtons = ['OK'];
+  public ngOnInit() {
+    this.credList = this.executionResponse['selectableVcList'];
+    this.credList.forEach(() => {
+      this.isClick.push(false);
+    });
+  }
 
-  setOpen(isOpen: boolean) {
+  public isClicked(index: number) {
+    return this.isClick[index];
+  }
+
+  public selectCred(cred: VerifiableCredential, index: number) {
+    this.selCredList.push(cred);
+    this.isClick[index] = !this.isClick[index];
+  }
+  public async sendCred(cred: VerifiableCredential) {
+    const alert = await this.alertController.create({
+      header: this.translate.instant('confirmation.header'),
+      buttons: [
+        {
+          text: this.translate.instant('confirmation.cancel'),
+          role: 'cancel',
+        },
+        {
+          text: this.translate.instant('confirmation.ok'),
+          role: 'ok',
+        },
+      ],
+    });
+
+    await alert.present();
+    const result = await alert.onDidDismiss();
+    console.log(result);
+
+    if (result.role === 'ok') {
+      this.selCredList.push(cred);
+      this._VCReply.selectedVcList = this.selCredList;
+      this.walletService.executeVC(this._VCReply).subscribe({
+        next: () => {
+          this.isAlertOpen = true;
+        },
+        error: (err) => {
+          console.error(err);
+        },
+        complete: () => {
+          this.selCredList = [];
+        },
+      });
+    }
+  }
+
+  public setOpen(isOpen: boolean) {
     this.isAlertOpen = isOpen;
   }
 }
