@@ -1,33 +1,49 @@
 import { TestBed } from '@angular/core/testing';
-import { HttpClientTestingModule } from '@angular/common/http/testing';
 import { AuthenticationService } from './authentication.service';
 import { OidcSecurityService } from 'angular-auth-oidc-client';
-import { StorageService } from './storage.service';
+import { of } from 'rxjs';
 
 describe('AuthenticationService', () => {
-  let authenticationService: AuthenticationService;
+  let service: AuthenticationService;
+  let oidcSecurityService: jasmine.SpyObj<OidcSecurityService>;
 
   beforeEach(() => {
-    const authServiceSpy = jasmine.createSpyObj('AuthenticationService', [true, '', ''])
-    const oidcSecuritySpy = jasmine.createSpyObj('OidcSecurityService', ['checkAuth', 'authorizeWithPopUp', 'logoff']);
-    const storageServiceSpyObj = jasmine.createSpyObj('StorageService', ['get', 'set', 'remove']);
+    oidcSecurityService = jasmine.createSpyObj('OidcSecurityService', ['checkAuth', 'authorizeWithPopUp', 'logoff']);
+
+    oidcSecurityService.checkAuth.and.returnValue(of({
+      isAuthenticated: true,
+      userData: { name: 'John Doe' },
+      accessToken: 'dummy-token',
+      idToken: 'dummy-id-token'
+    }));
 
     TestBed.configureTestingModule({
-      imports: [HttpClientTestingModule],
       providers: [
-        {provide: AuthenticationService, useValue: authServiceSpy},
-        { provide: OidcSecurityService, useValue: oidcSecuritySpy },
-        { provide: StorageService, useValue: storageServiceSpyObj },
-      ],
+        AuthenticationService,
+        { provide: OidcSecurityService, useValue: oidcSecurityService }
+      ]
     });
 
-    authenticationService = TestBed.inject(AuthenticationService);
-
+    service = TestBed.inject(AuthenticationService);
   });
 
   it('should be created', () => {
-    expect(authenticationService).toBeTruthy();
+    expect(service).toBeTruthy();
   });
 
+  it('checkAuth should complete without errors', (done) => {
+    oidcSecurityService.checkAuth.and.returnValue(of({ isAuthenticated: true, userData: {}, accessToken: '', idToken: '' }));
 
+    service.checkAuth().subscribe({
+      next: () => {},
+      error: () => fail('checkAuth should not have failed'),
+      complete: () => done()
+    });
+  });
+
+  it('logout should call oidcSecurityService.logoff', () => {
+    service.logout();
+    expect(oidcSecurityService.logoff).toHaveBeenCalled();
+  });
 });
+
