@@ -1,29 +1,35 @@
-import { Injectable} from '@angular/core';
-import { OidcSecurityService} from 'angular-auth-oidc-client';
-
+import { Injectable } from '@angular/core';
+import { OidcSecurityService } from 'angular-auth-oidc-client';
+import { BehaviorSubject, Observable, map } from 'rxjs';
 
 @Injectable({
   providedIn: 'root',
 })
 export class AuthenticationService {
-
-
+  public name: BehaviorSubject<string> = new BehaviorSubject<string>('');
   private token!: string;
+  // eslint-disable-next-line @typescript-eslint/no-explicit-any
   private userData: any;
 
-  constructor(public oidcSecurityService: OidcSecurityService) {
-    this.oidcSecurityService.checkAuth().subscribe(({ userData, accessToken }) => {
-      this.userData = userData;
-      this.token = accessToken;
-    });
+  public constructor(public oidcSecurityService: OidcSecurityService) {
+    this.checkAuth().subscribe();
+  }
+  public checkAuth() {
+    return this.oidcSecurityService.checkAuth().pipe(
+      map(({ userData, accessToken }) => {
+        this.userData = userData;
+        this.name.next(this.userData.name);
+        this.token = accessToken;
+      })
+    );
   }
   public logout() {
     return this.oidcSecurityService.logoff();
   }
-  public getToken():string{
-    return  this.token;
+  public getToken(): string {
+    return this.token;
   }
-  public getName():string {
-    return this.userData.preferred_username;
+  public getName(): Observable<string> {
+    return this.name;
   }
 }
