@@ -4,32 +4,29 @@ import { BehaviorSubject, catchError, map } from 'rxjs';
 import { environment } from 'src/environments/environment';
 
 @Injectable({
-  providedIn: 'root'
+  providedIn: 'root',
 })
 export class DataService {
-  private ebsiFlag: boolean = false;
-  private did: string = '';
-  private didSubject = new BehaviorSubject<string>("");
+  private didSubject = new BehaviorSubject<string>('');
 
-  constructor(private http: HttpClient) { }
+  public constructor(private http: HttpClient) {}
 
-  sendDid(did: string) {
-    this.didSubject.next(did);
-  }
-
-  listenDid(): any {
+  public listenDid(): BehaviorSubject<string> {
     return this.didSubject;
   }
 
-  getDid() {
-    return this.http.get(environment.server_url + environment.server_uri.ebsi_did_uri, { responseType: 'text' }).pipe(
-      map((data) => {
-        return data.toString()
-      }),
-      catchError((err) => {
-        throw new Error("Err");
+  public getDid() {
+    return this.http
+      .get(environment.server_url + environment.server_uri.ebsi_did_uri, {
+        responseType: 'text',
       })
-    )
+      .pipe(
+        map((didResponse) => {
+          this.didSubject.next(didResponse);
+        }),
+        catchError(() => {
+          throw new Error('Err');
+        })
+      );
   }
-
 }
