@@ -2,37 +2,45 @@ import { Injectable } from '@angular/core';
 import { AuthenticationService } from './authentication.service';
 import { BehaviorSubject } from 'rxjs';
 import { AlertController } from '@ionic/angular';
+import { environment } from 'src/environments/environment';
 
 @Injectable({
-  providedIn: 'root'
+  providedIn: 'root',
 })
 export class WebsocketService {
   private socket!: WebSocket;
-  private messageSubject = new BehaviorSubject<string>("");
+  private messageSubject = new BehaviorSubject<string>('');
 
-  constructor(  private authenticationService: AuthenticationService,
+  public constructor(
+    private authenticationService: AuthenticationService,
     private alertController: AlertController
+  ) {}
 
-    ) {}
-
-  connect(url: string): void {
-    this.socket = new WebSocket(url);
+  public connect(): void {
+    this.socket = new WebSocket(
+      environment.websocket_url + environment.websocket_uri
+    );
 
     this.socket.onopen = () => {
       console.log('Conexión WebSocket abierta');
-      this.sendMessage(JSON.stringify({ id: this.authenticationService.getToken() }));
+      this.sendMessage(
+        JSON.stringify({ id: this.authenticationService.getToken() })
+      );
     };
 
     this.socket.onmessage = async (event) => {
       console.log('Mensaje recibido:', event.data);
-
       const alert = await this.alertController.create({
         header: 'Introducir PIN',
         inputs: [
           {
             name: 'pin',
-            type: 'number',
+            type: 'text',
             placeholder: 'PIN',
+            attributes: {
+              inputmode: 'numeric',
+              pattern: '[0-9]*',
+            },
           },
         ],
         buttons: [
@@ -48,11 +56,10 @@ export class WebsocketService {
           },
         ],
       });
-  
+
       await alert.present();
       const result = await alert.onDidDismiss();
-       this.sendMessage(JSON.stringify({ pin: result.data?.values.pin || '' }));
-
+      this.sendMessage(JSON.stringify({ pin: result.data?.values.pin || '' }));
     };
 
     this.socket.onclose = () => {
@@ -60,7 +67,7 @@ export class WebsocketService {
     };
   }
 
-  sendMessage(message: string): void {
+  public sendMessage(message: string): void {
     if (this.socket.readyState === WebSocket.OPEN) {
       this.socket.send(message);
     } else {
@@ -68,7 +75,7 @@ export class WebsocketService {
     }
   }
 
-  closeConnection(): void {
+  public closeConnection(): void {
     this.socket.close();
   }
 }
