@@ -16,7 +16,9 @@ export const environment = {
     credentials_uri: window["env"]["credentials_uri"] || '/api/v2/credentials',
     credentials_by_id_uri: window["env"]["credentials_by_id_uri"] || '/api/v2/credentials?credentialId=',
     users_uri: window["env"]["users_uri"] || '/api/v2/users',
-    ebsi_did_uri: window["env"]["ebsi_did_uri"] || '/api/v2/ebsi-did'
+    ebsi_did_uri: window["env"]["ebsi_did_uri"] || '/api/v2/ebsi-did',
+    cbor: window["env"]["cbor_uri"] || '/api/v2/vp/cbor'
+
   },
   websocket_uri: window["env"]["websocket_uri"] || '/api/v2/pin'
 };
